refactor: Move Character class and related enums to Shared library
