--- Shared/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace project_axiom_sharelib;
-
-public class Class1
-{
-
-}
